:sparkles: Fixes #110 -- implement retrieving the form step definition
The API endpoint renders the representation of a particular formstep
in the context of a given submission. If no submission step exists yet
(because there was no data submitted yet), the read action still
returns the necessary information for the frontend to be able to
render the form.

Access control validates that the submission is active in your
current session, and requests are throttled by default to prevent
abuse.

Custom field type registration works and is able to transform Formio
components.