:goal_net: Refs. #152 -- handle case without recipient
In the event that the form field is optional or just completely absent,
do not crash hard but log a warning that the confirmation e-mail could
not be sent.

This will probably happen when people forget to specify the appropriate
field and the default is used, or there is just no e-mail in the form
at all.