:lipstick: Fix broken admin index styles
The global bootstrap reset/includes of the form builder mess with the
django admin styles (and django admin index styles), so we need
more specific CSS rules.

Now the admin index hover effect is no longer behind the breadcrumbs,
the dropdown menu and breadcrumbs have the correct height and the
colors are fixed.