:lock: Fixes #408 -- set session cookie SameSite flag to none
A middleware is added to set the SameSite flag explicitly to none, to
make sure that the session cookie is sent from cross-domain contexts.
Django < 3.1 does not support this out of the box, but this is needed
since the SDK will operate in cross-domain contexts.

The CSRF cookie remains SameSite=Lax (the default) and will therefore
not be set or sent in a cross-domain context, thereby still protecting
against actual CSRF attacks.

An finally, for the cross-site requests based on session cookie, the
CORS settings work with an allow-list, preventing attacks from evil.com
to the Open Forms submissions.

Note that this DOES NOT WORK on Chrom(e)(ium) 80+ on localhost, since
it enforces `SameSite=none;Secure` and we don't have SSL in the dev-
server or static HTML pages. Firefox emits a warning for this, so it
can still be tested there. On the test/staging/production environment,
everything is deployed on https and should be fine.