:bug: Fixes #449 -- fix toggling React CRUD feature flag
Bug caused by ModelAdmin instances not being thread safe. Mutating an
instance (by doing self.change_form_template) affects the next
requests as well, even if the feature flag says it should be otherwise.

Removed the thread-unsafe code and replaced it by passing the config
into the template and dealing with the different situations in the
template.

Side-effect: the SDK snippets are back. Not great yet, but they're back.