:alien: Specify type of custom components and pre-fill tab
The type must be supplied, and it must be a custom type, otherwise the form builder
uses the parent class type for the `editForm\' which is not always appropriate. The
logic for this can be seen here: https://github.com/formio/formio.js/blob/4.13.x/src/WebformBuilder.js\#L1255

Additionally, the pre-fill tab has been added for the Postcode and BSN fields, as those
are typically pre-fillable from a logged-in user.