:heavy_plus_sign: Refs #504 -- Added celery once
Celery Once ensures only a single instance of a task is running. It
guards against tasks running longer than the scheduled beat interval
which would cause a cascade of load.

By default, locks expire after an hour so that if something goes wrong,
it takes maximum one hour for the next job to be triggered again.