:ok_hand: Refs. #181 -- add 'close' button taking user back to original page
Each cookie group has its own set of accept/decline buttons, so the
logical redirect is to show the cookie list page again. However, that
resets the HTTP_REFERER headers so we can't rely on that anymore.

The solution is to move this parameter to the querystring and prefer
the qs value if it's present, falling back to the header if it's not.