:sparkles: #506 -- style form page header
The 'subdomain' approach styling for a form detail page is imlemented using
design tokens.

The design tokens allow control of the padding, logo size and even the logo
size itself. Furthermore, the administrators can configure the return URL
and upload the logo to use. The logo is clickable and returns you to the
main website URL (if configured). If no logo is configured, a regular link
will be displayed to return to the main website.