:bug: Refs. #651 -- mark tasks that shouldn't store the result in the redis backend
Using a result backend gives access to the meta information of the
task and the eventual result. This is stored as key-value pairs in
redis (the result backend). We should only do this for the tasks that
we consult (status check, result is used in other places...) to not
leak resources, since tasks that store the result should get reaped
via AsyncResult.get|forget.

The tasks which are fire-and-forget are now explicitly marked to not
store the result in the backend, so we don't need to track their
task IDs and afterwards clean up the resources.