:recycle: Fixes #839 -- replace registration backend retries with full on_completion_retry chain
The automatic retry would only retry sending the submission to the
registration backend if it had failed, but not execute downstream
tasks such as updating the appointment and/or payment status.

The celery beat task now queues the entire retry chain, and the
configuration parameters (environment variables) have been renamed
and updated accordingly.