:sparkles: Fixes #839 -- include zaak kenmerken if an existing public reference exists
If zaak-creation (StUF-ZDS or ZGW) fails, the on_completion chain will
generate an internal reference which is communicated to the end-user
who filled out the form.

The retry flow will then attempt (and hopefully succeed) in creating
the zaak, but we cannot guarantee an unique zaak-identificatie. We
can also not change the reference for the end-user, as this would be
a terrible user experience.

Zaken API and StUF-ZDS account for this by including 'kenmerken' which
can contain references to the same zaak in other registrations. We
include the internally generated reference in there if it exists.