:bug: Fixes #839 -- implement generic appointment create failure error message
In the event that the appointment cannot be created in the backend system because of a technical
issue (=not because of missing user data), store a generic error message to be displayed to the
end-user.

Additionally, the on_completion workflow should exit then and there, as we have nothing tangible
for the end-user.