:card_file_box: Refs. #235 -- add Submission.price field
Since product prices can change after a submission was completed, it's
important to freeze/record the price at the moment on the submission
itself. It's also possible the price is a calculated result from
logic rules, where the logic rules on the form may change or be removed
after the submission is completed.