:sparkles: [#326] -- ensure that object keys snake_case/camelCase is kept
Note that this applies _globally_ to the FormViewSet parsers and
renderers, and is expected to cause problems at some point. We leverage
the built-in drf-camel-case 'ignore_fields' option, but that may break
if other, non-camunda related backends define a DictField with the
'definition' key where keys _should_ be transformed.

There are a couple of notes/warnings/todos in the comments to generalize
this properly, and a future attempt at fine-grained json paths for
ignore_fields will probably be made and suggested to the upstream
library.