:bug: [#1174] -- fix appointment config modifications breaking
The root cause of this is the refactor in the getFormComponents
utility in 9e53f2a07, which constructs a new component object to
add the stepLabel key. Directly modifying the component in that call
is not possible because of immutability reasons.

In turn, this caused the form appointments configuration tab to break,
since that implicitly relied on the mutability of the Formio
components to set appointment-related properties.

This is now fixed in the reducer by mutating only draft state through
a new utility called findComponent.