:lock: Change the NUM_PROXIES default from None to 1
This assumes a single reverse proxy, such as nginx on simple servers
or an ingress controller on Kubernetes. This is a potentially breaking
change for existing installations if there are more reverse proxies
involved.

However, if the setting is left to 'None', clients can possibly bypass
API throttling by spoofing the relevant header.