:wrench: Fixes #1182 -- do not create a pidfile for celery beat
Normally, celery beat cleans up the pidfile on exit (even if it
crashes), but there are code paths with python stdlib crashes that
apparently aren't expected.

This crash happens when the broker/redis server goes away while beat
is running - which would typically happen during a reboot or deploy
when redis is restarted. On startup, beat starts gracefully with
exponential backoff if the connection is not available (yet), so that's
not the cause.

This patch instructs beat to NOT use a pidfile, which should not have
impact on normal operation. On crashes/restarts, beat should
succesfully restart since there's no pidfile left in the modified
container filesystem.