:bug: Fixes #1222 -- correct visual state when deleting logic rules
The UI now correctly renders the state when you delete a logic rule
other than the last one.

This was caused by React's key prop functionality combined with a bad
choice for the key value. React does not re-render a component if the
key has not changed, even if other props have changed (!).

Logic rules used their Array index as key value, so unless you delete
the last logic rule, the visual state will update incorrectly.

Deleting a logic rule removes it from the the Array - causing all
following rules to get an index value of  - 1. From
React's perspective, the keys have not changed - only there is now one
element less in the Array so the last rule is instead removed.
The internal state (for API calls) was correctly updated, however the
visual representation of this state was wrong.

This is fixed by using an actual unique value for each rule as key (
rule.uuid) if it's available, while falling back to a temporary
generated unique string for new rules that don't have a UUID yet. This
results in a guaranteed unique key value for every logic rule and the
correct React re-renders in turn.