:ok_hand: [#1193] -- swap order of delete operations
First perform the database delete, then clean up the filesystem afterwards.

The idea is that potential bugs resulting in broken queries do not lead to an
inconsistent state with files that are suddenly missing while the transaction is
rolled back (if a transaction is active).

Before file delete errors were captured and logged, it was preferable to keep the
records around with maybe missing files to be able to investigate, as this exception
would block the delete query from executing. However, with any exception caught while
attempting to delete the file, the query will always execute anyway. Therefor, it's
safer to schedule the file deletion after the DB query.

The drawback is that files may be orphaned and left on the file system while the
matching DB record has been deleted. However, the logging ensures that this does
happen in a visible way, and the filepath is now also logged to manually clean up
or investigate.

This also opens up the path to transaction.on_commit behaviour.