:bug: #1199 -- catch OIDC database errors and display them on custom error page
User creation may cause integrity errors based on email claims that
already exist for another user (may be the same actual user, but
Open Forms doesn't know that). Errors are now wrapped in an atomic
block and caught for display.

Additionally, we redirect to a failure URL, as refreshing the callback
URL with the code won't work anyway (it protects against replay
attacks).