:bug: Fixes #1399 -- streamline KVK (prefill) integration
The KVK exposes different types of 'companies' in its API, manifesting
in different blobs of information being returned (essentially the
resources are polymorphic without being explicitly modeled that way
in the API).

For example, a B.V. has a hoofdvestiging, but a VVE does not.

The endpoint consulted for company information is adapted to fetch
the base-profile instead of the hoofdvestiging (which may not exist
for certain types of companies). The response contains the hoofd-
vestiging (if it exists) in the _embedded container (per HAL spec).

It seems that either the hoofdvestiging or the 'eigenaar' embedded
resource contains the addresses, if any are present at all, so our
post-processing extracts those on a best-effort basis.

Tests have been adapted for the base-profile responses, and some
kvk.nl test-environment data added for the VVE-case.

The test to check access paths from the constants is removed, as this
was just a copy of the schema anyway and did not reflect a real-world
scenario where keys can be missing - none of the schema properties
are marked as required so we have no guarantees.