:bug: [#1422] Prevent update of custom keys on label changes
The issue is that when a component in the editForm with key 'label' is updated, the key of the main
component is reset. For Radio buttons / selectboxes, there are multiple fields in the editForm with key
'label': the label of the component and the labels of the options. From within the
this.editForm.everyComponent function it didn't seem possible to differentiate which component was being
edited. So, to prevent changes to the options labels to reset the key, a custom option "isOptionLabel"
was added to them in src/openforms/js/components/form/edit/tabs.js.