:bug: [#1466] -- handle non-string option values (radio, select,...)
While options for selects and radio inputs are defined and stored as
strings (for the value aspect), the SDK does funky things and guesses
the datatype of a value in 'auto' mode (unless the datatype is
explicitly specified). This results in the type of submission value
not matching the type of the option definition value (string) and
lead to assertion errors.

Instead, we now cast the value to string if it's not a string to find
a proper match.