:test_tube: Add regression test for order race condition
This is extremely hard to actually reproduce... I'm sorry.

Basically the idea is that multiple requests can arrive at the same
time (and they will do because of how the frontend is coded!) which kicks
off a number of threads that all start with the same view on the data.

django-ordered-model then performs some UPDATE queries based on that
view of the data and finally a save call to the object itself to
set the order. That last action often corrects a lot of things,
but the UPDATE query causes mayhem.

See also django-ordered-model/django-ordered-model#184.