:bug: Fixes #1611 -- use temporary step identifier for step select if step is not persisted
Instead of the empty URL, we now use the step._generatedId as a unique
reference for steps that only live in browser memory (so far). After saving the
steps to the backend, these steps are updated with their UUID and URL references,
making it possible to grab the actual URL from the temporary ID during the
saving phase.