:alien: [#1708] -- deal with localized numbers
Ideally, we want to pass unlocalized numbers in the formio configuration (dot
for decimal separator rather than comma), but that's not possible in
combination with the floatformat filter in Django 3.2. Newer Django versions
allow specifying a suffix for unlocalized formatting.

Formio seems to handle the localized values well though, as it's running in the NL locale.