:recycle: [#1708] -- properly leverage MessageList component
PR feedback pointed out some duplicated component code, but this was
not identical to the specified ErrorList component. When trying to
sort this out, it became apparent that the MessageList component could
be structured better and used appropriately.

The MessageList component now exists at the 'admin' level rather than
the form-designer level and all usage has been replaced.