:recycle: [#1913] -- refactored evaluating logic rules
Logic rule evaluation (including the processing of the actions) is now
refactored into a single function. The function takes care of updating the
variable state while processing the actions, and yields all other action/
operations to the caller who's responsible for processing them.

This centralizes the machinery to process a logic rule.