:children_crossing: [#1921] -- provide usable file type selection widget in admin
The default arrayfield with text inputs is not user friendly, as it
requires you type in mime type patterns. Using dropdowns instead
doesn't work with django-better-admin-arrayfield, as it has hardcoded
javascript looking for an input element rather than a select element.

Instead, we use the MultiChoiceField from Django with the relevant
choices.