:zap: [#1922] -- Optimize the bulk form variables endpoints
**Variables list**

* Removed unnecessary prefetch/select_related from variables-list
  endpoint
* Added select_related on variables-list endpoint for variable
  form and form-definition
* Variables-list endpoint now has a fixed number of queries: 2 (+
  session/middleware queries etc.)

**Variables bulk update**

Every form variable was being validated individually, meaning that a FK
lookup was performed for the form and the (optional) form definition,
giving a complexity somewhere between O(n) and O(2*n) with n the number
of form variables.

The form validation/lookup is not needed in the first place, since the
endpoint is scoped within a form detail endpoint. The form definition
validation is a bit more complex, but again, multiple variables can
belong to the same form-definition, so doing this query multiple times
is wasted performance.

The performance patch collects the form and it's related form
definitions upfront and passes them via the serializer context down to a
custom field which replaces the DB lookups with context lookups.