:zap: [#2042] -- implement a thread-local request-scoped proxy cache
This cache backend is intended to be used with django-solo and the
first cache layer, leveraging the Python dict access/memory. On
cache misses/writes, the cache calls are proxied to the configured
underlying cache (which would be redis in our case).

The cache is only activated within a request-response cycle, and
cleared when a request is finished, which solves the problem of cache
invalidation when the solo instance is changed in another process or
thread.

This primarly eliminates the multitude of network calls to Redis within
a single request-response cycle, which has considerable network overhead
and latency.