:sparkles: [#1979] -- implement upgrade path check before migrating
Added checks to verify if an upgrade path is possible as part of the
system check framework. If an upgrade path is not possible, this returns
an error, which in turn blocks the 'migrate' command (due to failing
system checks).

In other situations where the from/to version cannot be reliably
determined or compared, a warning is emitted. In development, no
errors or warnings are shown at all.