:fire: [#2145] -- remove copy-paste snippets from form change page
The generated snippets are dynamic in nature - they typically require a
proper CSP nonce and also require the correct URL path inclusion. This
makes it that they're not easily copy-pasted and saved into a CMS
somewhere, as the snippet itself needs updating if the page location
changes.

Additionally, typically a catch-all route needs to be configured in the
third party wishing to embed the form so that client-side routing does
not break - the SDK does not yet offer the option to use hash-fragment
based routing.

Rather than showing a 'broken' or incorrect snippet, we opt to not show
it all anymore and instead refer to the embed/integration documentation.