:recycle: [#1068] -- refactor custom formio field types to use dynamic config approach
Now it's just another flavour of mutating an existing component on
the fly, with the added functionality of making some network calls to
retrieve the relevant data.

This essentially does away with the whole special treatment of custom
field types. The use case of adding one or more components and
rebuilding the component tree is removed, as there has not been any
demand for this.

We do keep in mind it may at some point be needed to inject components
at other locations in the tree or do more advanced operations, but
the current implementation does not sit in the way of that approach.

Additionally, this fixes the bug where only top-level npFamilyMembers
components could function correctly, which is now properly recursive
for nested components.