:lock: Addressed some minor security concerns
The bootstrap JS does not need to be loaded on error templates (and it
probably doesn't work either since no jQuery is loaded), so it was
removed.

The inline-styles gained the request csp nonce to allow them to be
applied. Note that the bootstrap styles themselves are still blocked
by the CSP, so we need to make a decision here.