:recycle: [#2418] -- Refactor translation handling in client after backend rework
We no longer need to deal with form.literals or formStep.literals, as
these fields are exposed via form.translations and formStep.translations
with a more straight-forward data model.

FormStep/FormDefinition translations are still managed in a single
object structure, which is merged/split as needed while saving/loading
the data to/from the backend.

Currently there are a couple of places that display or use the 'name'
of the form/steps, which relies on the top-level key (which is
translated according to the active language). That can be refactored
at a later stage.