:bug: [#2418] -- ensure template validation errors are plain text
There's only a benefit of getting the template syntax with proper
formatting if DEBUG=true, which is never the case in the situations
where a form-designer/administrator is editing templates and getting
validation errors. This makes the whole displaying-the-template-source
to point out syntax errors obsolete.

Instead, we emit plain text errors now that should contain sufficient
information, which is suitable for DRF serialization, as the SafeString
details of rendered-template-as-validation-error are lost when DRF
processes Django's validation errors. There is also no suitable
mechanism to communicate from backend to frontend that a certain
validation error contains (sanitized) HTML and can be rendered in an
un-safe manner (with XSS risks as result).