:bug: [#2418] -- Fix label-click-toggles-disabled-checkbox because ID was not unique
The label 'for' attribute refers to the ID of the HTML input element.
The name and ID of the disabled checkboxes in translation tabs was
duplicated, because the value is language-independent (which is why
we disable it for the non-default language codes in the first place).

Because the ID was duplicated, the label was linked to the first
occurrence of the checkbox - the non-disabled one.

This patch unsets ID/htmlFor for disabled checkboxes and radios.