:recycle: [#2315] Process componentTranslations right after saving
before, the componentTranslations were stored in `state.formSteps[x]`, but this approach makes deletion of existing translations difficult, because we are comparing a change to a single component to the merged list of componentTranslations.

Now, this process is done right before actually submitting the form. The translations are constructed from the FormIO configuration and the translation references are deleted from the configuration