:bug: [#2484] Tweak fallbacks for modeltranslations
fallbacks based on default values cause issues in the case of literals, for instance
when a Form uses a global literal default (e.g. `Begin form`). In that scenario,
modeltranslations would find that default and fallback to a language that does
not use a default value for this literal, which is not what we want.

In the literal getter we have a manual fallback procedure, so we disable
the django-mt fallbacks temporarily while accessing the form/step
specific field(s).