✅[#2286] Test SubmissionStep translations
Assumptions:
 1. Changing language mid submission abandons it and starts a fresh one,
    so Submission.language_code == _active locale_
 2. All Submitter/End User visible text that needs to be translated
    lives in the `formio.variables.SUPPORTED_TEMPLATE_PROPERTIES`

Under these assumptions, all strings in components as they are served to
the client will already have been translated. And no extra
"translations" object as asked in #2286 is actually needed.

Therefore we just test that interpolation is done on the strings in the
submission locale in:
 - submission step endpoint
 - _check_logic endpoint

NB assumption 2 is false: in the form designer one could set properties
like "prefix", "suffix", "tooltip" by using the "editJson" button and
expect these to be picked up by translation.

And there may be strings that *are* exposed form designer that aren't
passed through template rendering. These could either be added, or
`inject_variable` could me made to iterate over a superset of
`SUPPORTED_TEMPLATE_PROPERTIES`, say user_visible_properties, and always translate,
but only `render` if the property is in `SUPPORTED_TEMPLATE_PROPERTIES`.