:fire: [#2555] -- remove code related to 2.0.0 upgrade check
The minimum required version for the next release will be 2.0.0, which
still has all these checks in that release cycle.

The checks are not executed again (anymore) when upgrading to 2.1, and
if they report anythying, there are bigger issues anyway.

This means the management commands and supporting scripts are no longer
relevant and can be safely removed.