🐛[#2475] Remove unused literals before PUT/POST
When a source language literal is changed by editing a component, a new
entry is added to each componentTranslation

  "new literal": "original translation"

and leaves the

  "original literal": "original translation"

in tact. This is correct as "original literal" might be used in a
different component in this step definition!
So just before posting/putting, when we have a full view of all literals
in use, we filter.