:bug: Fixes open-formulieren/security-issues#20 -- use fully qualified URLs in analytics config
Since the analytics config manages the CSP directive configuration,
it's imperative that the https:// protocol is included in the base URLs
of the analytics providers.

This requires changes to the templates that load the analytics scripts,
where the 'same-protocol' directive ('//') is replaced with the value
from the database, making it explicit.

Existing installations are advised to update their configuration and
review the CSP configuration at /admin/config/cspsetting/.