🩹[#2546] Add submission language to registration email content
This patch made me realise that the Content-Language header is a bit of
a lie: the message is a mixture of submission.language_code and
settings.LANGUAGE_CODE. 🤷🙈

I still think submission.language_code is the most useful for routing
purposes and processing rules. Getting the mail in the right [folder of
the ] mailbox of the right "medewerker".