:bug: [#2581] -- handle possible string inputs for date-like values
Discussed in office - essentially we have two options:
* crashing hard
* silently casting to None, removing any expected constraints from the
  date picker, possibly allowing garbage into registration backends that was
  not anticipated

We picked option 1, as the end-user will receive a generic error message from
the SDK, so they can contact their municipality, who in turn contact the service
provider, who in turn can look up the crash/error in Sentry and provide the
fixes.

Option 2 is harder to debug without visbility in error monitoring and therefore
not desired.