⚡[#2454] Reduce test time from ~50s to ~5s
`build_client()` would try to re-parse the OAS yaml file on every
example because the `Client` dataclass would have `_schema` set to
`None`.

`seek(0)` was a band-aid to allow it to `read()` again. But it's waaay
cheaper to just return the same client if we are called with the same
`**claims`.

Don't know if that whole `_schema` business upstream is doing what we
expect it to do...

Still need to raise the default deadline to silence hypothesis flakiness
warning when the all suites are run.
First example for first test run will sometimes take more than 200ms.