:recycle: [#388] -- Clean up namespaces/attributes
The 'main' or sector namespace is removed from the client and base
context. The SOAP envelope defines the soap/StUF and XSI namespaces,
and it's up to child templates to define their local sector namespaces
as they are relevant.

This patch also drops the xsi:schemaLocation attributes - the URLs and/
or filepaths in the values did not resolve anywhere, making the whole
attribute obsolete in terms of functionality.