:pushpin: [#2748] -- Pin to Python 3.10.9
It appears that Python 3.10.10 introduced a regression that can cause
deadlocks on threading.local which manifests in CI at the moment for us.

For "better safe than sorry" reasons the Docker image version is also
pinned as we do NOT want a risk of deadlocks in production.

For more information, see https://github.com/python/cpython/issues/102126