:bug: [#2842] Include analytics url in default-src directive
Before, the URL for the analytics tool was being included in CSP directives such as connect-src or font-src. Since there was no default value specified for these directives, requests to the open-forms backend were being blocked. Now we include the URL of the analytics tool in the default-src directive which contains 'self'. In this way requests to OF are not blocked. All other CSP directives default back to the value of default-src if not specified

Backport-Of: open-formulieren/open-forms#2843