:bug: Fix key not being updated in the modal
I suspect that there's a race condition happening with some mutation of the
literals ref and the form builder being forced to re-render in combination
with the useOnChanged hook. This then possibly leads to the modal generating
the key again from the label instead of using the uniqified key?

As far as I can tell right now, removing this doesn't seem to cause
problems, but you never know with this fragile setup...