:wastebasket: Deprecate the password field
After checking customer instances and getting in touch with customers,
it turns out nobody is actually using this component. Given that the
data is stored in plain text and pre-existing security/expectation
concerns, it is decided to not invest any more effort but instead move
the component to the deprecated group.

In the future we'll probably convert any possible existing components
to textfields and remove it alltogether.