:recycle: Refine the intended behaviour of the AllOrNoneTruthyFieldsValidator
The DRF definition of 'empty' does not align with our functional
definition of 'empty' - the intended usage of the validator is to ensure
that usable values are set for all fields and not a subset of fields, or
no values are set at all.

Usable values here are non-empty strings, values for choice fields other
than the empty/no choice...

Note that this does not work as expected when using boolean fields as
it'll require them to be True ({"foo": False, "bar": "baz"}) will not
validate.