🐛[#2911] Assume image/heif and image/heic are the same.
heif is a container format. heic is a heif file with a certain type of
compressed file in there.

libmagic doesn't recognize all heics as such, nor can PIL/pillow
in dialogue with the product owner, this solution suffices for the
usecase.

Alternative would be to do all of the following:
- add heif explicitly to the allowed filetype options which isn't great
  UX;
- install extra dependency (like wand or a pillow file reader) to inspect the
  whole file.