♻[#2893] Rip out "context"
So apparently this is true in `get_input_components`:
`assert input_data == context`

Even when context is a made deepcopy of `data_container.data`, though
the tests may not be exercising all side-effect code paths while
asserting the log output.

I remember not seeing "today" in there, but I guess I must have been looking wrong.
Yet the name `input_data` suggests strongly that it shouldn't be there;
"today" is not "input_data" by the form submitter agent (human or GPT).

Our caller passes in `initial_data` as our parameter `input_data`, which
suggests it's data from before... ... ...

The whole exercise of backporting the patches that add the
action_log_data to EvaluatedRule could be rolled back. But the tests
pass, so they might be harmless. Bottom line is that there is *some*
data logged, and no JSON in triggers or values should be able to crash
the task anymore.