:sparkles: [#2957] -- enhance the design tokens JSON field with editor
Added a button and pretty JSON formatting for the raw admin text field.
Clicking the button opens the design token editor in a modal where
users get a nicer interface to manage the tokens. By default, our
curated tokens are displayed (via @open-formulieren/design-tokens
package and metadata in the tokens themselves).

Power users still have access to the raw JSON field in the admin and can
copy-paste and/or edit the resulting JSON.