✨[#2959] Add union type Either, implement algorithm W
The union type Either is used as the return type of the "if" operator as
it may have the type of the then-clause or it may have the type of the
else-clause. Some components may accept optional values (Either Null
SomeType) so forcing then and else clauses to have the same type hampers
expressiveness too much.

This implements W and switches type_check to using W. Because M unifies
"as-it-goes" unification the type of a in Either a a is not inferred
yet. Because W infers all types before unifying, Either a a can be
simplified to a easily.

W does give broader (harder to understand) error when expressions don't
type check. M pinpoints the error location to narrowest sub-expression
within the expression, that doesn't check.

For M to handle Either, the way we need it, it will need changes to
postpone the unification of Either applications untill both parameters
have a known type, but before M returns (!) or it will return
intermediate typevars (like [t0]) as the inferred type.