:white_check_mark: [#2918] -- updated E2E integration tests
The existing tests are subclasses so that we can test both the variants with
feature flag disabled (default) and enabled. This required some parametrization
in the tests due to slightly different HTML input names (but they are
functionally equivalent).

Once the feature flag is gone, the tests can be simplified again.