:green_heart: Fix race condition in test
The save of the form configuration happens async with a bunch of API calls. So,
the JS is still processing by the time the test advances to the assertion stage,
and depending on whether the form step transaction was already committed or not,
the test can fail.

Fixed by adding an explicit assertion to check/wait for redirection back
to the form list, which guarantees all save transactions have committed.