:bug: [#2651] -- fix analytics cookies receiving incorrect domain
The domain was populated from settings.ALLOWED_HOSTS[0], which is not
sufficiently deterministic nor intended for this behaviour.

Cookie domain information is now taken from settings.BASE_URL and the
setting is added to the documentation as required configuration. It
was already used in such a way (e.g. when building absolute URLs
without HttpRequest instance available), documentation was just lacking
for this setting.

Additionally, the domain hash input had the same problem, so that's
been rectified in the same manner.