🐛[#2860] Support unencrypted SAML Assertions
I think this is a bug that hasn't surfaced. We checked all our known
deployments, and they all "Want assertions encrypted" in their configs.

SAML Artefact encryption and signing comes from the (old) flow where all
communication is done through the requesting browser. So the assertion
needs to be tamper proof (signed) and may contain private data that
needs to be protected (encrypted).

We (and most brokers) only support the Artefact Binding flow, where the
artefact is acquired over a back channel with the broker/IdP.